Changes after merge and review
