+++ tests/parsers/csv_parser.py
@@ -1,14 +1,10 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
-"""Tests for the comma seperated values (CSV) parser."""
+"""Tests for the comma separated values (CSV) parser."""
 
 from __future__ import unicode_literals
 
 import unittest
-
-from dfvfs.file_io import os_file_io
-from dfvfs.path import os_path_spec
-from dfvfs.resolver import context
 
 from plaso.parsers import csv_parser
 
@@ -59,102 +55,6 @@ class TestCSVParser(csv_parser.CSVParser):
     return True
 
 
-@shared_test_lib.skipUnlessHasTestFile(['another_file'])
-@shared_test_lib.skipUnlessHasTestFile(['password.txt'])
-class BinaryLineReaderTest(shared_test_lib.BaseTestCase):
-  """Tests for the binary line reader."""
-
-  def setUp(self):
-    """Sets up the needed objects used throughout the test."""
-    self._resolver_context = context.Context()
-    test_file = self._GetTestFilePath(['another_file'])
-    self._os_path_spec1 = os_path_spec.OSPathSpec(location=test_file)
-
-    test_file = self._GetTestFilePath(['password.txt'])
-    self._os_path_spec2 = os_path_spec.OSPathSpec(location=test_file)
-
-  def testReadline(self):
-    """Test the readline() function."""
-    file_object = os_file_io.OSFile(self._resolver_context)
-    file_object.open(self._os_path_spec1)
-    line_reader = csv_parser.BinaryLineReader(file_object)
-
-    self.assertEqual(line_reader.readline(), b'This is another file.\n')
-
-    self.assertEqual(line_reader.get_offset(), 22)
-
-    file_object.close()
-
-  def testReadlines(self):
-    """Test the readlines() function."""
-    file_object = os_file_io.OSFile(self._resolver_context)
-    file_object.open(self._os_path_spec2)
-    line_reader = csv_parser.BinaryLineReader(file_object)
-
-    lines = line_reader.readlines()
-
-    self.assertEqual(len(lines), 5)
-    self.assertEqual(lines[0], b'place,user,password\n')
-    self.assertEqual(lines[1], b'bank,joesmith,superrich\n')
-    self.assertEqual(lines[2], b'alarm system,-,1234\n')
-    self.assertEqual(lines[3], b'treasure chest,-,1111\n')
-    self.assertEqual(lines[4], b'uber secret laire,admin,admin\n')
-
-    file_object.close()
-
-  def testReadlinesWithSizeHint(self):
-    """Test the readlines() function."""
-    file_object = os_file_io.OSFile(self._resolver_context)
-    file_object.open(self._os_path_spec2)
-    line_reader = csv_parser.BinaryLineReader(file_object)
-
-    lines = line_reader.readlines(sizehint=60)
-
-    self.assertEqual(len(lines), 3)
-    self.assertEqual(lines[0], b'place,user,password\n')
-    self.assertEqual(lines[1], b'bank,joesmith,superrich\n')
-    self.assertEqual(lines[2], b'alarm system,-,1234\n')
-
-    file_object.close()
-
-  @shared_test_lib.skipUnlessHasTestFile(['mactime.body'])
-  def testReadlinesWithFileWithoutNewLineAtEnd(self):
-    """Test reading lines from a file without a new line char at the end."""
-    test_file = self._GetTestFilePath(['mactime.body'])
-    test_file_path_spec = os_path_spec.OSPathSpec(location=test_file)
-    file_object = os_file_io.OSFile(self._resolver_context)
-    file_object.open(test_file_path_spec)
-    line_reader = csv_parser.BinaryLineReader(file_object)
-
-    lines = line_reader.readlines()
-
-    self.assertEqual(len(lines), 17)
-
-  def testIterator(self):
-    """Test the iterator functionality."""
-    file_object = os_file_io.OSFile(self._resolver_context)
-    file_object.open(self._os_path_spec2)
-    line_reader = csv_parser.BinaryLineReader(file_object)
-
-    lines = []
-    for line in line_reader:
-      lines.append(line)
-
-    self.assertEqual(len(lines), 5)
-    self.assertEqual(lines[0], b'place,user,password\n')
-    self.assertEqual(lines[1], b'bank,joesmith,superrich\n')
-    self.assertEqual(lines[2], b'alarm system,-,1234\n')
-    self.assertEqual(lines[3], b'treasure chest,-,1111\n')
-    self.assertEqual(lines[4], b'uber secret laire,admin,admin\n')
-
-    file_object.close()
-
-  # TODO: Add a test which tests reading a file which is
-  # larger than the buffer size, and read lines until it crosses
-  # that original buffer size (to test if the buffer is correctly
-  # filled).
-
-
 class CSVParserTest(test_lib.ParserTestCase):
   """Tests the comma seperated values (CSV) parser."""
 
@@ -188,12 +88,12 @@ class CSVParserTest(test_lib.ParserTestCase):
     unicode_row = parser._ConvertRowToUnicode(parser_mediator, binary_row)
     self.assertEqual(unicode_row['password'], 'superr\xedch')
 
-  @shared_test_lib.skipUnlessHasTestFile(['password.txt'])
+  @shared_test_lib.skipUnlessHasTestFile(['password.csv'])
   def testParseFileObject(self):
     """Tests the ParseFileObject function."""
     parser = TestCSVParser()
 
-    self._ParseFile(['password.txt'], parser)
+    self._ParseFile(['password.csv'], parser)
 
     self.assertEqual(len(parser.rows), 4)
