+++ tests/parsers/mcafeeav.py
@@ -27,7 +27,9 @@ class McafeeAccessProtectionUnitTest(test_lib.ParserTestCase):
     # The file contains 14 lines which results in 14 events.
     self.assertEqual(storage_writer.number_of_events, 14)
 
-    events = list(storage_writer.GetEvents())
+    # The order in which CSVParser generates events is nondeterministic
+    # hence we sort the events.
+    events = list(storage_writer.GetSortedEvents())
 
     event = events[0]
