+++ tests/parsers/symantec.py
@@ -45,7 +45,9 @@ class SymantecAccessProtectionUnitTest(test_lib.ParserTestCase):
     # The file contains 8 lines which should result in 8 events.
     self.assertEqual(storage_writer.number_of_events, 8)
 
-    events = list(storage_writer.GetEvents())
+    # The order in which CSVParser generates events is nondeterministic
+    # hence we sort the events.
+    events = list(storage_writer.GetSortedEvents())
 
     # Test the second entry:
     event = events[1]
